Refactor to use math.isfinite
Previously we used math.isnan, math.isinf, and a custom _isan function in various places.
Now that we support infinity in the input, we can use math.isfinite instead.

This commit refactors the _convert_to_float function to be near the caller, guess_series.
We also convert the bare except clause to an Exception.